ui-tester/requirements.txt
volcengine-python-sdk[ark]>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
